hf_dashboard/requirements.txt
gradio>=4.44.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pandas>=2.2.0
openpyxl>=3.1.2
jinja2>=3.1.3
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
Pillow>=10.2.0
email-validator>=2.1.0
python-dateutil>=2.8.2
python-dotenv>=1.0.1
requests>=2.31.0
